+++ docs/normalize_notebooks.py
@@ -1,6 +1,0 @@
-#!/usr/bin/env python
-"""
-Normalize Jupyter notebooks
-"""
-import docutils as ut
-ut.normalize_notebooks()