+++ starsim/calibration.py
@@ -18,7 +18,9 @@ op = sc.importbyname('optuna', lazy=True)
 sns = sc.importbyname('seaborn', lazy=True)
 vis = sc.importbyname('optuna.visualization.matplotlib', lazy=True)
 
-__all__ = ['Calibration']
+# Defaults
+default_n_boot = 1000 # Default number of bootstrap samples for plot_facet_bootstrap()
+
 
 
 _sim_pool_data = {} # This is used to minimize pickling to workers, mainly beneficial on Windows and Mac OS. There will be one local copy on each Optuna worker
@@ -626,9 +628,7 @@ class Calibration(sc.prettyobj):
 
 #%% Calibration components
 
-__all__ += ['linear_interp', 'linear_accum', 'step_containing'] # Conformers
-__all__ += ['CalibComponent'] # Calib component base class
-__all__ += ['BetaBinomial', 'Binomial', 'DirichletMultinomial', 'GammaPoisson', 'Normal'] # Specific calib components
+
 
 def linear_interp(expected, actual):
     """
@@ -770,6 +770,7 @@ class CalibComponent(sc.prettyobj):
         return conform_
 
     def _combine_reps_nll(self, expected, actual, **kwargs):
+        """ Aggregate replicates (if combine_reps is set) and compute negative log likelihood """
         if self.combine_reps is None:
             nll = self.compute_nll(expected, actual, **kwargs) # Negative log likelihood
         else:
@@ -918,7 +919,7 @@ class BetaBinomial(CalibComponent):
         expected = self.expected.loc[t]
         e_n, e_x = expected['n'], expected['x']
 
-        n_boot = kwargs.get('n_boot', 1000)
+        n_boot = kwargs.get('n_boot', default_n_boot)
         seeds = data['rand_seed'].unique()
         boot_size = len(seeds)
         means = np.zeros(n_boot)
@@ -1000,7 +1001,7 @@ class Binomial(CalibComponent):
         expected = self.expected.loc[t]
         e_n, e_x = expected['n'], expected['x']
 
-        n_boot = kwargs.get('n_boot', 1000)
+        n_boot = kwargs.get('n_boot', default_n_boot)
         seeds = data['rand_seed'].unique()
         boot_size = len(seeds)
         means = np.zeros(n_boot)
@@ -1174,7 +1175,7 @@ class GammaPoisson(CalibComponent):
         expected = self.expected.loc[[t]]
         e_n, e_x = expected['n'].values.flatten()[0], expected['x'].values.flatten()[0]
 
-        n_boot = kwargs.get('n_boot', 1000)
+        n_boot = kwargs.get('n_boot', default_n_boot)
         seeds = data['rand_seed'].unique()
         boot_size = len(seeds)
         means = np.zeros(n_boot)
@@ -1292,7 +1293,7 @@ class Normal(CalibComponent):
         expected = self.expected.loc[[t]] # Gracefully handle Series and DataFrame, if 't1' in index
         e_x = expected['x'].values.flatten()[0] # Due to possible presence of 't1' in the index
 
-        n_boot = kwargs.get('n_boot', 1000)
+        n_boot = kwargs.get('n_boot', default_n_boot)
         seeds = data['rand_seed'].unique()
         boot_size = len(seeds)
         means = np.zeros(n_boot)